pyproject.toml
[project]
name = "linearsteering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.5.0",
    "loguru>=0.7.3",
    "torch>=2.9.1",
    "transformer-lens>=2.16.1",
    "matplotlib>=3.10.0",
    "umap-learn>=0.5.9.post2",
    "ipykernel>=7.1.0",
    "python-dotenv>=1.0.1",
    "openai>=2.17.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.3.4",
    "ruff>=0.9.4",
]
